template: Update to new interface
