+++ apps/govuk-template/package.json
@@ -37,6 +37,7 @@
     "@not-govuk/anchor-list": "workspace:^0.4.0",
     "@not-govuk/app-composer": "workspace:^0.4.0",
     "@not-govuk/app-plop-pack": "workspace:^0.4.0",
+    "@not-govuk/asset-proxy": "workspace:^0.4.2",
     "@not-govuk/client-renderer": "workspace:^0.4.0",
     "@not-govuk/components": "workspace:^0.4.0",
     "@not-govuk/engine": "workspace:^0.4.0",
